+++ client/src/components/Navbar/ProfileDrawer.vue
@@ -1,10 +1,17 @@
 <template>
-  <v-navigation-drawer v-model="ProfileDrawerState" absolute temporary right @input="showProfileDrawer($event)" 
-  :stateless="ProfileModalState || ChangePasswordModalState">
+  <v-navigation-drawer
+    v-model="ProfileDrawerState"
+    absolute
+    temporary
+    right
+    @input="showProfileDrawer($event)"
+    :stateless="ProfileModalState || ChangePasswordModalState"
+  >
     <template v-slot:prepend>
       <v-list-item two-line>
         <v-list-item-avatar>
-          <img src="https://randomuser.me/api/portraits/women/81.jpg" />
+          <!-- <img src="https://randomuser.me/api/portraits/women/81.jpg" /> -->
+          <img src="https://picsum.photos/1920/1080?random" />
         </v-list-item-avatar>
 
         <v-list-item-content>
@@ -15,8 +22,7 @@
     </template>
     <v-divider></v-divider>
     <v-list dense>
-
-      <v-list-item link>
+      <v-list-item link @click="reload">
         <v-list-item-icon>
           <v-icon>mdi-home-city</v-icon>
         </v-list-item-icon>
@@ -51,7 +57,6 @@
           <v-list-item-title>Logout</v-list-item-title>
         </v-list-item-content>
       </v-list-item>
-
     </v-list>
   </v-navigation-drawer>
 </template>
@@ -67,25 +72,55 @@ export default {
     showDrawer() {
       this.show = !this.show;
     },
-    showProfileModal(){
-      this.$store.commit('showProfileModal')
+    reload() {
+      window.location.reload();
     },
-    showChangePasswordModal(){
-       this.$store.commit('showChangePasswordModal')
+    showProfileModal() {
+      this.$store.commit("showProfileModal");
     },
-    showProfileDrawer(e){
-      if (!e){
-        this.$store.commit('showProfileDrawer')
+    showChangePasswordModal() {
+      this.$store.commit("showChangePasswordModal");
+    },
+    showProfileDrawer(e) {
+      if (!e) {
+        this.$store.commit("showProfileDrawer");
       }
     },
     async Logout() {
       try {
-        let result = await AccountService.Logout();
-        console.log(result);
-        window.sessionStorage.clear();
-        this.$router.go();
+        let value = await this.$swal({
+          title: "Confirm To Logout ?",
+          icon: "question",
+          showCancelButton: true,
+          confirmButtonText: "Yes",
+          cancelButtonText: "Cancel",
+          showCloseButton: true,
+        });
+        if (value.isConfirmed) {
+          let result = await AccountService.Logout();
+          console.log(result);
+          this.$swal({
+            title: "Logout Successfully",
+            icon: "success",
+            timer: 1000,
+            showConfirmButton: false,
+            showCancelButton: false,
+          }).then((result) => {
+            if (result.dismiss) {
+              window.sessionStorage.clear();
+              this.$router.go();
+            }
+          });
+        }
       } catch (err) {
         console.log(err.response);
+        this.$swal({
+          title: "Logout Fail, Please try again",
+          icon: "error",
+          timer: 1000,
+          showConfirmButton: false,
+          showCancelButton: false,
+        });
       }
     },
   },
@@ -93,30 +128,30 @@ export default {
     user() {
       return this.$store.getters.getUser;
     },
-    ProfileDrawerState:{
-      get:function() {
-        return this.$store.getters.getProfileDrawer
+    ProfileDrawerState: {
+      get: function() {
+        return this.$store.getters.getProfileDrawer;
       },
-      set:function(newValue){
-        return newValue
-      }
+      set: function(newValue) {
+        return newValue;
+      },
     },
-    ProfileModalState:{
-      get:function() {
-        return this.$store.getters.getProfileModal
+    ProfileModalState: {
+      get: function() {
+        return this.$store.getters.getProfileModal;
       },
-      set:function(newValue){
-        return newValue
-      }
+      set: function(newValue) {
+        return newValue;
+      },
     },
-    ChangePasswordModalState:{
-      get:function() {
-        return this.$store.getters.getChangePasswordModal
+    ChangePasswordModalState: {
+      get: function() {
+        return this.$store.getters.getChangePasswordModal;
       },
-      set:function(newValue){
-        return newValue
-      }
-    }
+      set: function(newValue) {
+        return newValue;
+      },
+    },
   },
   created() {
     this.$root.$refs.ProfileDrawer = this;
